Organize errors module
