+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/method/errors/mod.rs
@@ -1,0 +1,2 @@
+mod invalid_class_member;
+mod invalid_class_prop;
